Solve case when there are no senders
What happens when all the senders are dropped and the reciever is
waiting? Should it wait infinitely? It should not.

First, Our current Inner struct is re-named into shared, and creating a
new inner struct which also maintains a senders field, to count the
number of senders, and this is updated whenever clone or drop occurs on
Sender. Whenever a drop occurs and the sender see's that its the last
one it also performs a notify (which was only done for send). and the
reciever while recieving checks if senders are 0, and returns a None in
that case as Recieve is -> Option<T> now.

The vice-versa case i.e no recievers but senders are present is not that
important as senders don't block unlike recievers and can be ignored as
the program does not wait.

Signed-off-by: Tarun Pothulapati <tarunpothulapati@outlook.com>